checked if file exists

+++ src/D2L.Bmx/WriteHandler.cs
@@ -70,10 +70,12 @@ internal class WriteHandler(
 			if( !data.Sections.ContainsSection( sectionName ) ) {
 				data.Sections.AddSection( sectionName );
 			}
-			var defaultCredentialsFile = parser.ReadFile( SharedCredentialsFile.DefaultFilePath );
-			if (defaultCredentialsFile.Sections.ContainsSection( profile ) ) {
-				defaultCredentialsFile.Sections.RemoveSection( profile );
-				parser.WriteFile(SharedCredentialsFile.DefaultFilePath, defaultCredentialsFile);
+			if (File.Exists(SharedCredentialsFile.DefaultFilePath)) {
+				var defaultCredentialsFile = parser.ReadFile( SharedCredentialsFile.DefaultFilePath );
+				if ( defaultCredentialsFile.Sections.ContainsSection( profile ) ) {
+					defaultCredentialsFile.Sections.RemoveSection( profile );
+					parser.WriteFile( SharedCredentialsFile.DefaultFilePath, defaultCredentialsFile );
+				}
 			}
 			data[sectionName]["credential_process"] =
 				"bmx print --format json --cache --non-interactive"
